backend/requirements.txt
asgiref==3.8.1
Django==5.1
django-environ==0.11.2
psycopg2==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.1
djangorestframework
requests
bs4
pandas
django-cors-headers